+++ src/outbound.ts
@@ -4,6 +4,7 @@ import { getActiveClient } from "./monitor.js";
 import { bareJid, resolveServer } from "./config-schema.js";
 import { getUploadService, uploadAndGetUrl, buildOobElement, downloadUrl } from "./http-upload.js";
 import { isOmemoEnabled, encryptOmemoMessage, encryptMucOmemoMessage, buildOmemoMessageStanza, isRoomOmemoCapable } from "./omemo/index.js";
+import { sentMessageIds } from "./state.js";
 
 export interface ResolvedMedia {
   data: Buffer;